Decided to allow users to create as many rusults as they have memory.

+++ FizzBuzzGenerator/Extensions.cs
@@ -4,7 +4,7 @@ namespace FizzBuzzGenerator
 {
     public static class Extensions
     {
-        public static bool IsDivisibleBy(this Int16 i, short divisor)
+        public static bool IsDivisibleBy(this int i, int divisor)
         {
             return i % divisor == 0;
         }
